+++ sdk/compute/Azure.ResourceManager.Compute/tests/Samples/Readme_ManageAvailabilitySet.cs
@@ -97,7 +97,7 @@ namespace Azure.ResourceManager.Compute.Tests.Samples
             string rgName = "myRgName";
             ResourceGroup resourceGroup = await rgCollection.GetAsync(rgName);
             string availabilitySetName = "myAvailabilitySet";
-            bool exists = await resourceGroup.GetAvailabilitySets().CheckIfExistsAsync(availabilitySetName);
+            bool exists = await resourceGroup.GetAvailabilitySets().ExistsAsync(availabilitySetName);
 
             if (exists)
             {