+++ sdk/confidentialledger/Azure.Security.ConfidentialLedger/src/ConfidentialLedgerClient.cs
@@ -38,10 +38,10 @@ namespace Azure.Security.ConfidentialLedger
             var authPolicy = new BearerTokenAuthenticationPolicy(_tokenCredential, AuthorizationScopes);
             _pipeline = HttpPipelineBuilder.Build(
                 actualOptions,
-                new HttpPipelinePolicy[] { new LowLevelCallbackPolicy() },
+                Array.Empty<HttpPipelinePolicy>(),
                 new HttpPipelinePolicy[] { authPolicy },
-                new ResponseClassifier(),
-                transportOptions);
+                transportOptions,
+                new ResponseClassifier());
             _ledgerUri = ledgerUri;
             _apiVersion = actualOptions.Version;
         }