+++ sdk/eventhub/Azure.ResourceManager.EventHubs/tests/Samples/Readme_ManagingNamespaces.cs
@@ -76,7 +76,7 @@ namespace Azure.ResourceManager.EventHubs.Tests.Samples
             {
                 Console.WriteLine("namespace 'foo' exists");
             }
-            if (await namespaceCollection.CheckIfExistsAsync("bar"))
+            if (await namespaceCollection.ExistsAsync("bar"))
             {
                 Console.WriteLine("namespace 'bar' exists");
             }