+++ sdk/resourcemanager/Azure.ResourceManager/tests/Scenario/ProviderOperationsTests.cs
@@ -18,7 +18,7 @@ namespace Azure.ResourceManager.Tests
         public void NoDataValidation()
         {
             ///subscriptions/db1ab6f0-4769-4b27-930e-01e2ef9c123c/providers/microsoft.insights
-            var resource = Client.GetProvider($"/subscriptions/{Guid.NewGuid()}/providers/microsoft.FakeNamespace");
+            var resource = Client.GetProvider(new ResourceIdentifier($"/subscriptions/{Guid.NewGuid()}/providers/microsoft.FakeNamespace"));
             Assert.Throws<InvalidOperationException>(() => { var data = resource.Data; });
         }
 
@@ -31,7 +31,8 @@ namespace Azure.ResourceManager.Tests
             Provider result = response.Value;
             Assert.IsNotNull(result);
 
-            var ex = Assert.ThrowsAsync<RequestFailedException>(async () => await Client.GetProvider(result.Data.Id + "x").GetAsync());
+            ResourceIdentifier fakeId = new ResourceIdentifier(result.Data.Id.ToString() + "x");
+            var ex = Assert.ThrowsAsync<RequestFailedException>(async () => await Client.GetProvider(new ResourceIdentifier(fakeId)).GetAsync());
             Assert.AreEqual(404, ex.Status);
         }