+++ sdk/resourcemanager/Azure.ResourceManager/tests/Unit/ResourceListOperationsTest.cs
@@ -15,9 +15,9 @@ namespace Azure.ResourceManager.Tests
             return new Plan("name", "publisher", "product", "promo", "version");
         }
 
-        private static Sku GetSku()
+        private static Resources.Models.Sku GetSku()
         {
-            return new Sku("name", SkuTier.Basic, "family", "size", 10);
+            return new Resources.Models.Sku("name", SkuTier.Basic.ToString(), "size", "family", "model", 10);
         }
 
         private static GenericResourceData GetGenericResource()
@@ -33,13 +33,13 @@ namespace Azure.ResourceManager.Tests
 
         private static GenericResourceData GetGenericResource(
             Dictionary<string, string> tags,
-            Sku sku,
+            Resources.Models.Sku sku,
             Plan plan,
             string kind,
             string managedBy,
             string location)
         {
-            ResourceIdentifier id = $"/subscriptions/{Guid.NewGuid().ToString()}/resourceGroups/myResourceGroup/providers/Microsoft.Widgets/widgets/myWidget";
+            ResourceIdentifier id = new ResourceIdentifier($"/subscriptions/{Guid.NewGuid()}/resourceGroups/myResourceGroup/providers/Microsoft.Widgets/widgets/myWidget");
             return new GenericResourceData(id, id.Name, id.ResourceType, location, tags, plan, null, kind, managedBy, sku, null, null, null, null);
         }
     }