+++ sdk/websites/Azure.ResourceManager.AppService/tests/TestsCase/StaticSiteCollectionTests.cs
@@ -71,17 +71,17 @@ namespace Azure.ResourceManager.AppService.Tests.TestsCase
         [TestCase]
         [RecordedTest]
         [Ignore("RepositoryToken is invalid.Please ensure the repository exists and the RepositoryToken is for an admin of the repository.")]
-        public async Task CheckIfExistsAsync()
+        public async Task Exists()
         {
             var container = await GetStaticSiteCollectionAsync();
             var staticSiteName = Recording.GenerateAssetName("testStaticSite");
             var input = ResourceDataHelper.GetBasicStaticSiteARMResourceData(StaticSiteLocation);
             var lro = await container.CreateOrUpdateAsync(staticSiteName, input);
             StaticSiteARMResource staticSiteARMResource = lro.Value;
-            Assert.IsTrue(await container.CheckIfExistsAsync(staticSiteName));
-            Assert.IsFalse(await container.CheckIfExistsAsync(staticSiteName + "1"));
+            Assert.IsTrue(await container.ExistsAsync(staticSiteName));
+            Assert.IsFalse(await container.ExistsAsync(staticSiteName + "1"));
 
-            Assert.ThrowsAsync<ArgumentNullException>(async () => _ = await container.CheckIfExistsAsync(null));
+            Assert.ThrowsAsync<ArgumentNullException>(async () => _ = await container.ExistsAsync(null));
         }
     }
 }